Most of the specified features work
Books are listed in 3 "shelves": 1) Currently Reading, 2) Want To Read, and 3) Finished Reading. Implemented the search form to allow a user to type in a term to search the book database. The found books are listed in 4 groups as the user types in a term: books not yet on a shelf and one group for each of the found books that are already on a shelf. A drop down menu is displayed with each of the books to allow the user to place the book in a shelf.