More modularity
Separated the code into more modules as recommended by the Udacity reviewer. Most of the other recommendations were also followed. Still need to need add nicer buttons to take the user from the Main Page to the Search Page. Was not able to reproduce the TypeError from calling this.state.searchBooks.filter from Apps.js so do not know if this is still present.